+++ templates/ai_help.html
@@ -19,7 +19,7 @@
             </p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_ai') }}"
+           href="{{ url_for('ai.view_ai') }}"
            title="Back to AI">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to AI</span>
         </a>
