+++ templates/data_management_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">TTL, backup, restore, and scheduler integration guidance with practical examples.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_dm_settings') }}"
+           href="{{ url_for('k8s_dm.view_dm_settings') }}"
            title="Back to Data Management">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Data Management</span>
         </a>
