+++ templates/kubernetes_help.html
@@ -19,7 +19,7 @@
             </p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_kubernetes') }}"
+           href="{{ url_for('k8s_dm.view_kubernetes') }}"
            title="Back to Health View">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Health View</span>
         </a>
