+++ templates/settings_help.html
@@ -92,11 +92,11 @@
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-info w-100"
-                       href="{{ url_for('view_k8s_settings') }}">Kubernetes</a>
+                       href="{{ url_for('k8s_dm.view_k8s_settings') }}">Kubernetes</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-danger w-100"
-                       href="{{ url_for('view_dm_settings') }}">Data Management</a>
+                       href="{{ url_for('k8s_dm.view_dm_settings') }}">Data Management</a>
                 </div>
             </div>
         </div>
