+++ templates/settings_kubernetes.html
@@ -28,7 +28,7 @@
     {% endif %}
     <div class="card border-secondary mb-4">
         <div class="card-body">
-            <form method="post" action="{{ url_for('save_k8s_settings') }}">
+            <form method="post" action="{{ url_for('k8s_dm.save_k8s_settings') }}">
                 <div class="mb-4">
                     <div class="form-check form-switch">
                         <input class="form-check-input"
@@ -49,7 +49,7 @@
                         <i class="bi bi-floppy me-1"></i>Save Settings
                     </button>
                     {% if k8s_settings.get('kubernetes.enabled') == '1' %}
-                        <a href="{{ url_for('view_kubernetes') }}"
+                        <a href="{{ url_for('k8s_dm.view_kubernetes') }}"
                            class="btn btn-sm btn-outline-info">
                             <i class="bi bi-hdd-network me-1"></i>Open Dashboard
                         </a>
