+++ templates/summary.html
@@ -123,7 +123,7 @@
     </a>
   </div>
   <div class="col-6 col-md-4 col-lg-2">
-     <a href="{{ url_for('view_ai') }}" class="text-decoration-none"
+     <a href="{{ url_for('ai.view_ai') }}" class="text-decoration-none"
        data-ai-action-id="summary.nav.ai"
        data-ai-action-type="navigate"
        data-ai-handler="navigate"
