+++ templates/work_items_help.html
@@ -133,7 +133,7 @@
                     <a href="{{ url_for('settings.view_agent_rules') }}">Agent Settings</a> — create and manage agent rules that trigger work item creation.
                 </li>
                 <li>
-                    <a href="{{ url_for('view_ai') }}">AI Transparency</a> — audit all AI and agent calls that led to work item creation.
+                    <a href="{{ url_for('ai.view_ai') }}">AI Transparency</a> — audit all AI and agent calls that led to work item creation.
                 </li>
             </ul>
         </div>
